Fix bug where player would not be assigned clues
In order to determine which players should receive clues, the function
`char_roles` in `src/gamedata.py` uses self.guild.roles and checks if
`role.members` exists. That is, we fetch all members of the guild (discord
server) that have the role for the character. However, without using the
"Server Members Intent" setting in the bot, and declaring the intent in
the code, `role.members` may not be populated correctly. Declaring this
intent causes discord to actually check all members of the guild for the
correct role.

Oddly, Discord seems to still populate `role.members` with incomplete
data even if the intent is not declared. It seems like everyone online
or in a voice channel will be included. This is why this code has
probably worked just fine for most people.

The use-case that exposed this bug is as follows: We had five players,
two of which were co-located physically. Both players had discord
accounts and claimed roles for the game. However, only one of them was
connected to the voice channel so that they could share a webcam and
mic. The other player was connected via a mobile device. I suspect that
this caused the player on mobile to not show up when running
`role.members`.