Fix !upload
The previous code was setting "shell=True" to try to use > for output
redirection, but was still passing a list of arguments. When "shell=True" is
set, only a single string is expected for the entire command and all arguments
(e.g. "curl -T ...  > out_file"), so subprocess was actually spawning "curl"
with no arguments at all. This failure was hidden because the default
"check=False" ignores the return code of the spawned process, which is a
non-zero value for curl when run with no arguments.

This new code removes the need for a temporary file and instead reads the
stdout of the subprocess directly. It also sets "check=True" so that if curl
exits with an error, it will manifest as an exception and get reported/logged.